feat: load an ODRL policy from n-triples file
Functionality to read an ODRL policy from an n-triples file, currently based on
a hardcoded filename.  The contents of the file will be parsed using
`cl-ntriples`, a parser for the n-triples format.  The triples objects from the
parser are converted to escaped strings and used as body in a SPARQL `INSERT`
query.