fix: convert each ODRL object in a policy exactly once
The ODRL-to-ACL conversion starts converting objects from the rules in a policy
and from there goes to the referenced asset and group collections to convert
them. Consequently, asset and group collections that are use in multiple rules
are visited multiple times. Resulting in multiple ACL objects for a single
collection.

To prevent this we pass the ACL configuration being created as state with the
conversion methods. Furthermore, the ACL is extended with functions to add new
entities in way that ensures no duplicate entities are added to a given configuration.