refactor[policy]: rewrote the functionality to retrieve policy data
Retrieve all data for a policy using a single `CONSTRUCT` query instead of using
multiple `SELECT` queries.  The received jsown objects are passed as argument to
the functions creating ODRL/SHACL object instances. The individual `make-*`
functions search for and parse the necessary data from that argument. This
avoids them querying the backend each time.

The order in which to ODRL/SHACL object instances are created is
inverted. Previously, the creation started from permissions and from there moved
on to the linked party collections, asset collections, and assets. But since
multiple permissions can link to the same target and/or assignee collection this
flow can cause multiple object instances to be created for the same resource.
To avoid this duplication, the order is now:
1) for each Asset resource, create a `shacl:node-shape` object;
2) for each Asset Collection and Party Collection resource, create an
   `odrl:asset-collection` or `odrl:party-collection` object respectively; and
3) for each Permission resource, create an `odrl:permission` object.

The result of each step is used as part of the input for the subsequent step. So
the `shacl:node-shape`s instances from step 1 are used when creating
`odrl:asset-collection`s in step 2, that way an asset collection can reference
to its contained assets. Similarly, step 3 takes the created
odrl:asset-collection`s and `odrl:party-collection`s from step 2 as input so
that the created `odrl:permission`s reference the appropriate targets and
assignees.