feat[policy]: raise errors if mandatory data is missing in a policy
Previously there were no explicit checks on whether all necessary data was
present when converting a policy. For example, each party collection *must* have
a name. Consequently, the service would handle data issues rather inelegantly
with crashes etc.

These changes make the service more resilient with respect to faulty and/or
incomplete policy data in the backend. Introduce a condition
`no-data-for-mandatory-slots-error` that can be used to signal when an object
instance cannot be created due missing data for a mandatory slot.

To simplify signalling that condition the `with-mandatory-slots` macro is
introduced. This macro allows to specify a list of mandatory slots and their
intended initialisation value. If any of the provided initialisation values is
nil a `no-data-for-mandatory-slots-error` will be signalled.

If a condition is signalled, a message informing the user of this is printed to
the standard output. In such as case no configuration file will be written, and
thus no existing files will be overwritten.