feat(devx): support for live reloading
If the environment variable `MODE` has as value "development", enable live
reloading to simplify developing this service.  Otherwise, launch a REPL as
usual.

In development mode, the REPL will be launched using the `run-development`
script, provided as command to `watchexec`.  Every time a change to a lisp, asd,
or lock file is detected, `watchexec` will signal a `SIGUSR1` to the script's
process and, after it exited, re-execute.  Thereby launching a new REPL that
includes the made changes.

As part of the `run-development` script the REPL process will take the place of
script's process upon launch.  Using the functionality provided by the
`trivial-signal` package, the REPL process keeps listing for the `SIGUSR1`
signal and will exit `sbcl` upon receiving it.