feat: support loading prefixes on boot and when loading policy data
Use the previously added functionality to load prefixes from a default file on
boot. This allows users to more easily "configure" the service with a set of
prefixes that is should always know.

Furthermore, when loading a policy from an N-triples files, check to see whether
there is a TTL file with the same name. If so, load the prefixes declared in
that file. The assumption is that if an N-triples and TTL file have the same
name, the former is likely generated from the latter.